Reduce query offload to worker path
